Recycling_Segmentation/Recycling_AI/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
safetensors
pillow
opencv-python-headless
scipy
numpy
scikit-image
